+++ CHANGELOG.md
@@ -1,9 +1,19 @@
-<!-- latest_release 1.29.9 -->
+<!-- latest_release 1.31.0 -->
+## [v1.31.0](https://github.com/inspec/inspec-aws/tree/v1.31.0) (2020-10-01)
+
+#### Merged Pull Requests
+- dynamodb table resource [#270](https://github.com/inspec/inspec-aws/pull/270) ([samscottchef](https://github.com/samscottchef))
+<!-- latest_release -->
+
+## [v1.30.0](https://github.com/inspec/inspec-aws/tree/v1.30.0) (2020-09-23)
+
+#### Merged Pull Requests
+- guardduty detector resource WIP: Do not merge - waiting new Inspec release [#260](https://github.com/inspec/inspec-aws/pull/260) ([samscottchef](https://github.com/samscottchef))
+
 ## [v1.29.9](https://github.com/inspec/inspec-aws/tree/v1.29.9) (2020-09-15)
 
 #### Merged Pull Requests
 - Remove duplicate terraform plan creation at setup_integration_tests [#263](https://github.com/inspec/inspec-aws/pull/263) ([omerdemirok](https://github.com/omerdemirok))
-<!-- latest_release -->
 
 ## [v1.29.8](https://github.com/inspec/inspec-aws/tree/v1.29.8) (2020-09-14)